+++ confmap/provider.go
@@ -41,9 +41,13 @@ type Provider interface {
 	//
 	// `uri` must follow the "<scheme>:<opaque_data>" format. This format is compatible
 	// with the URI definition (see https://datatracker.ietf.org/doc/html/rfc3986). The "<scheme>"
-	// must be always included in the `uri`. The scheme supported by any provider MUST be at
-	// least 2 characters long to avoid conflicting with a driver-letter identifier as specified
-	// in https://tools.ietf.org/id/draft-kerwin-file-scheme-07.html#syntax.
+	// must be always included in the `uri`. The "<scheme>" supported by any provider:
+	//   - MUST consist of a sequence of characters beginning with a letter and followed by any
+	//     combination of letters, digits, plus ("+"), period ("."), or hyphen ("-").
+	//     See https://datatracker.ietf.org/doc/html/rfc3986#section-3.1.
+	//   - MUST be at least 2 characters long to avoid conflicting with a driver-letter identifier as specified
+	//     in https://tools.ietf.org/id/draft-kerwin-file-scheme-07.html#syntax.
+	//   - For testing, all implementation MUST check that confmaptest.ValidateProviderScheme returns no error.
 	//
 	// `watcher` callback is called when the config changes. watcher may be called from
 	// a different go routine. After watcher is called Retrieved.Get should be called
@@ -53,7 +57,7 @@ type Provider interface {
 	//
 	// If ctx is cancelled should return immediately with an error.
 	// Should never be called concurrently with itself or with Shutdown.
-	Retrieve(ctx context.Context, uri string, watcher WatcherFunc) (Retrieved, error)
+	Retrieve(ctx context.Context, uri string, watcher WatcherFunc) (*Retrieved, error)
 
 	// Scheme returns the location scheme used by Retrieve.
 	Scheme() string
@@ -106,15 +110,15 @@ func WithRetrievedClose(closeFunc CloseFunc) RetrievedOption {
 //   - Primitives: int, int32, int64, float32, float64, bool, string;
 //   - []interface{};
 //   - map[string]interface{};
-func NewRetrieved(rawConf interface{}, opts ...RetrievedOption) (Retrieved, error) {
+func NewRetrieved(rawConf interface{}, opts ...RetrievedOption) (*Retrieved, error) {
 	if err := checkRawConfType(rawConf); err != nil {
-		return Retrieved{}, err
+		return nil, err
 	}
 	set := retrievedSettings{}
 	for _, opt := range opts {
 		opt(&set)
 	}
-	return Retrieved{rawConf: rawConf, closeFunc: set.closeFunc}, nil
+	return &Retrieved{rawConf: rawConf, closeFunc: set.closeFunc}, nil
 }
 
 // AsConf returns the retrieved configuration parsed as a Conf.